create 'should_add_deposit_line' unit skeleton

+++ src/main/java/kata/bank/Account.java
@@ -3,7 +3,7 @@ package kata.bank;
 import java.util.Date;
 
 public class Account {
-    public void deposit(int amount, Date date) {
-        
+    public void deposit(double amount, Date date) {
+
     }
 }
